Fix test issues
